Break out end_to_end tests
